Refactor: Replace useless numbers in test filenamess with descriptions
The expected XML/JSON filenames had an integer value which is printed
by JUnit as part of the MethodSource ParameterizedTest execution.
As a result these filenames are useless as the description of what
each filename meant was within the Java constants.

Inline each constant, and rename the expected filenames to be more
descriptive.